backend/requirements.txt
# VPN Backend - Production dependencies
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-environ>=0.11.2
psycopg[binary]>=3.1.0
redis>=5.0.0
django-redis>=5.4.0
gunicorn>=21.0.0
whitenoise>=6.6.0
cryptography>=41.0.0
django-cors-headers>=4.3.0
